move current route functions into controllers
